orthanc/requirements.txt
flask
requests
pydicom
numpy
Pillow
